HSK 3.0 Wiktionary Scraper Initial
The HSK standardized test released a new version, 3.0. The new version
listed words and pronunciation but not the meanings. Also, this was
released as a PDF. The words here are from an OCR read in another github
repository, grabbing only the words without pronunciation.

I take these words and try to find the first "good" definition I get on
Wiktionary. It's very basic and thus incomplete and has many errors, but
it gives a baseline of results.